Make Redis optional in cert_signer_library, to avoid having to setup redis in low-infra scenarios

+++ backend/certificate_signer_library/redis.js
@@ -3,27 +3,53 @@ const {promisify} = require("util");
 let client;
 let existsAsync;
 let redisKeyExpiry;
+let isRedisEnabled;
 
 async function initRedis(config) {
-  client = redis.createClient(config.REDIS_URL);
-  client.on("error", function (error) {
-    console.error(error);
-  });
-  existsAsync = promisify(client.exists).bind(client);
+  isRedisEnabled = config.REDIS_ENABLED;
+  if(isRedisEnabled) {
+    client = redis.createClient(config.REDIS_URL);
+    redisConnectionEventListeners({ conn: client });
+    existsAsync = promisify(client.exists).bind(client);
+  }
   redisKeyExpiry = config.REDIS_KEY_EXPIRE;
 }
 
+function redisConnectionEventListeners({ conn }) {
+  conn.on('connect', () => {
+    console.log('Redis - Connection status: connected');
+  });
+  conn.on('end', () => {
+    console.log('Redis - Connection status: disconnected');
+  });
+  conn.on('reconnecting', () => {
+    // console.log('Redis - Connection status: reconnecting');
+  });
+  conn.on('error', (err) => {
+      console.error('Redis - Connection status: error ', { err });
+  });
+}
+
 async function checkIfKeyExists(key) {
-  return existsAsync(key)
+  if(isRedisEnabled && client.connected) {
+    return existsAsync(key)
+  } else {
+    return false
+  }
 }
 
 function storeKeyWithExpiry(key, value, expiry = redisKeyExpiry) {
-  client.set(key, value, "EX", expiry)
+  if(isRedisEnabled && client.connected) {
+    client.set(key, value, "EX", expiry)
+  }
 }
 
 function deleteKey(key) {
-  client.del(key)
+  if(isRedisEnabled && client.connected) {
+    client.del(key)
+  }
 }
+
 module.exports = {
   checkIfKeyExists,
   storeKeyWithExpiry,
